split subscription plan
