+++ src/pages/users/my_account_page.tsx
@@ -73,6 +73,7 @@ const getStaticPropsFunction = async ({ locale }: ILocale) => ({
       'asset',
       'dashboard',
       'date_picker',
+      'subscriptions',
     ])),
     locale,
   },
